+++ src/Bozboz/Admin/Reports/Filters/ArrayListingFilter.php
@@ -1,16 +1,27 @@
 <?php namespace Bozboz\Admin\Reports\Filters;
 
 use Illuminate\Support\Facades\Form;
+use Illuminate\Support\Facades\Input;
 
 class ArrayListingFilter extends ListingFilter
 {
+	protected $options;
+
+	protected $default;
+
+	public function __construct($name, $options, $callback = null, $default = null)
+	{
+		parent::__construct($name, $callback);
+
+		$this->options = $options;
+		$this->default = $default;
+	}
+
 	protected function defaultFilter($field)
 	{
 		return function($builder, $value) use ($field)
 		{
-			if ($value) {
-				$builder->where($field, $value);
-			}
+			$builder->where($field, $value);
 		};
 	}
 
@@ -20,5 +31,10 @@ class ArrayListingFilter extends ListingFilter
 		$html .= Form::select($this->name, $this->options, null, ['onChange' => 'this.form.submit()', 'class' => 'form-control']);
 
 		return $html;
+	}
+
+	public function getValue()
+	{
+		return Input::get($this->name, $this->default);
 	}
 }