+++ src/migrations/2015_02_19_143128_add_meta_title_to_pages_table.php
@@ -1,0 +1,36 @@
+<?php
+
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class AddMetaTitleToPagesTable extends Migration {
+
+	/**
+	 * Run the migrations.
+	 *
+	 * @return void
+	 */
+	public function up()
+	{
+		Schema::table('pages', function(Blueprint $table)
+		{
+			$table->string('meta_title')->after('html_title');
+		});
+
+		DB::statement('UPDATE `pages` SET `meta_title` = `html_title`');
+	}
+
+	/**
+	 * Reverse the migrations.
+	 *
+	 * @return void
+	 */
+	public function down()
+	{
+		Schema::table('pages', function(Blueprint $table)
+		{
+			$table->dropColumn('meta_title');
+		});
+	}
+
+}