+++ changelog.md
@@ -1,3 +1,6 @@
+# Version 0.4.3 (2015-02-19)
+-   Rename "html_title" field to "meta_title"
+
 # Version 0.4.2 (2015-01-26)
 -   Remove "menu_icon" field from PageAdminDecorator
 -   Change private methods `getRedirectOptions` and `getTemplateOptions` to protected